% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{contest}
\alias{contest}
\title{Tournament and race contest games}
\usage{
contest(x, n, type = c("race", "tournament"), prize = c(1, 0),
  elasticity = c(-1, 2, -1), p = punif, d = dunif, xlim = c(0, 1),
  deadline = 1, target = 0.4, ...)
}
\arguments{
\item{x}{vector of ability values}

\item{n}{number of competitors}

\item{type}{a contest game can be either a `race` or a `tournament`}

\item{prize}{value of 1st and 2nd prize (must sum to one)}

\item{elasticity}{vector of length 3 with elasticities for the cost function with respect ability, performance, and time.}

\item{p, d}{ability distribution and density}

\item{xlim}{vector of lowest and highest possible abilities}

\item{deadline, target}{time deadline and minimum performance target}

\item{...}{further arguments for distribution and density}
}
\value{
A `contest` object
}
\description{
Computes equilibrium bids and expected utility of a contest game with `n` rivals competing for two prizes of value `prizes` for a vector of iid abilities `x` drawn from a given distribution `p` with density `d`.
}
\examples{
contest(seq(0, 1, length=10), n=10, type="race")
}
\keyword{contest,}
\keyword{game}
\keyword{races,}
\keyword{theory}
\keyword{tournaments,}
