% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{plot.contest}
\alias{plot.contest}
\title{Plot contest object}
\usage{
\method{plot}{contest}(x, ...)
}
\arguments{
\item{x}{contest object}

\item{...}{further arguments for distribution and density}
}
\value{
Probability of the k-th order statistic from n iid random variables
}
\description{
Plot equilibrium bids and expected utility of a contest game with `n` rivals competing for two prizes of value `prizes` for a vector of iid abilities `x` drawn from a given distribution `p` with density `d`.
}
\examples{
plot(example(contest))
}
\keyword{order}
\keyword{statistics}
