% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_parallel.R
\name{run_parallel}
\alias{run_parallel}
\title{Title Runs run_simulation in parallel over parameters 
Provides a wrapper to perform a line search of a parameter and returns the 
results of run_simmulation with the parameters specified.}
\usage{
run_parallel(params, varied_parameter, from, to, by, ...)
}
\arguments{
\item{params}{Parameters for `run_simulation`}

\item{varied_parameter}{}

\item{from}{Starting value of varied parameter, See `seq`.}

\item{to}{Final value of varied parameter. See `seq`.}

\item{by}{Increment value. See `seq`.}

\item{mc_opts}{Options for mclapply, setting mc.cores is a good idea.}
}
\value{
A list of the result of calling run_simulation with the varied parameter.
}
\description{
Title Runs run_simulation in parallel over parameters 
Provides a wrapper to perform a line search of a parameter and returns the 
results of run_simmulation with the parameters specified.
}
\examples{
 params <- list(max_gens = 100,
                males_per_winner = 20,
                num_nests = 50,
                density = 1,
                metabolism = 1,
                female_mat_time = 10, 
                maturation_rate = 1,
                mutation_rate = 0.1,
                mutation_sd = 0.1,
                mass_to_energy = 10,
                growth_a = 0.5,
                growth_b = 0.1,
                initial_mass = 5,
                alpha_mean = 0,
                alpha_sd = 3,
                beta_sd = 3,
                beta_max = 20,
                beta_mean = 0,
                verbose = F,
                quiet = T)
 run_parallel(params, "density", 0.1, 4, 0.5, mc.cores = 4) -> sim
 Reduce(rbind, sim) -> df
 plot(df$generation, df$alpha, col = df$density)
 
}

