% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_from_json.R
\name{run_sim_from_json}
\alias{run_sim_from_json}
\title{Runs a simulation using parameters defined in a Json file}
\usage{
run_sim_from_json(file)
}
\arguments{
\item{file}{The .json file defining the parameters.}
}
\value{
The result of calling run_simulation
}
\description{
Runs a simulation using parameters defined in a Json file
}
\examples{
run_sim_from_json('parameters.json')
}

